+++ src/handlers/meta_transaction_handlers.ts
@@ -8,6 +8,7 @@ import * as isValidUUID from 'uuid-validate';
 
 import { CHAIN_ID } from '../config';
 import { DEFAULT_QUOTE_SLIPPAGE_PERCENTAGE, META_TRANSACTION_DOCS_URL } from '../constants';
+import { TransactionEntity } from '../entities';
 import {
     GeneralErrorCodes,
     generalErrorCodeToReason,
@@ -22,7 +23,7 @@ import { logger } from '../logger';
 import { isAPIError, isRevertError } from '../middleware/error_handling';
 import { schemas } from '../schemas/schemas';
 import { MetaTransactionService } from '../services/meta_transaction_service';
-import { GetTransactionRequestParams } from '../types';
+import { GetTransactionRequestParams, ZeroExTransactionWithoutDomain } from '../types';
 import { parseUtils } from '../utils/parse_utils';
 import { schemaUtils } from '../utils/schema_utils';
 import { findTokenAddressOrThrowApiError } from '../utils/token_metadata_utils';
@@ -190,7 +191,83 @@ export class MetaTransactionHandlers {
             throw new InternalServerError(e.message);
         }
     }
+    public async submitZeroExTransactionIfWhitelistedAsync(req: express.Request, res: express.Response): Promise<void> {
+        const apiKey = req.header('0x-api-key');
+        if (apiKey !== undefined && !isValidUUID(apiKey)) {
+            res.status(HttpStatus.BAD_REQUEST).send({
+                code: GeneralErrorCodes.InvalidAPIKey,
+                reason: generalErrorCodeToReason[GeneralErrorCodes.InvalidAPIKey],
+            });
+            return;
+        }
+        schemaUtils.validateSchema(req.body, schemas.metaTransactionFillRequestSchema);
 
+        // parse the request body
+        const { zeroExTransaction, signature } = parsePostTransactionRequestBody(req);
+        const zeroExTransactionHash = await this._metaTransactionService.getZeroExTransactionHashFromZeroExTransactionAsync(
+            zeroExTransaction,
+        );
+        const transactionInDatabase = await this._metaTransactionService.findTransactionByHashAsync(
+            zeroExTransactionHash,
+        );
+        if (transactionInDatabase !== undefined) {
+            // user attemps to submit a transaction already present in the database
+            res.status(HttpStatus.OK).send(marshallTransactionEntity(transactionInDatabase));
+            return;
+        }
+        try {
+            const protocolFee = await this._metaTransactionService.validateZeroExTransactionFillAsync(
+                zeroExTransaction,
+                signature,
+            );
+
+            // If eligible for free txn relay, submit it, otherwise, return unsigned Ethereum txn
+            if (apiKey !== undefined && MetaTransactionService.isEligibleForFreeMetaTxn(apiKey)) {
+                const {
+                    ethereumTransactionHash,
+                    signedEthereumTransaction,
+                } = await this._metaTransactionService.submitZeroExTransactionAsync(
+                    zeroExTransactionHash,
+                    zeroExTransaction,
+                    signature,
+                    protocolFee,
+                );
+                // return the transactionReceipt
+                res.status(HttpStatus.OK).send({
+                    ethereumTransactionHash,
+                    signedEthereumTransaction,
+                });
+            } else {
+                const ethereumTxn = await this._metaTransactionService.generatePartialExecuteTransactionEthereumTransactionAsync(
+                    zeroExTransaction,
+                    signature,
+                    protocolFee,
+                );
+                res.status(HttpStatus.FORBIDDEN).send({
+                    code: GeneralErrorCodes.UnableToSubmitOnBehalfOfTaker,
+                    reason: generalErrorCodeToReason[GeneralErrorCodes.UnableToSubmitOnBehalfOfTaker],
+                    ethereumTransaction: {
+                        data: ethereumTxn.data,
+                        gasPrice: ethereumTxn.gasPrice,
+                        gas: ethereumTxn.gas,
+                        value: ethereumTxn.value,
+                        to: ethereumTxn.to,
+                    },
+                });
+            }
+        } catch (e) {
+            // If this is already a transformed error then just re-throw
+            if (isAPIError(e)) {
+                throw e;
+            }
+            // Wrap a Revert error as an API revert error
+            if (isRevertError(e)) {
+                throw new RevertAPIError(e);
+            }
+            logger.info('Uncaught error', e);
+            throw new InternalServerError(e.message);
+        }
+    }
     public async getTransactionStatusAsync(req: express.Request, res: express.Response): Promise<void> {
         const transactionHash = req.params.txHash;
         try {
@@ -208,15 +285,7 @@ export class MetaTransactionHandlers {
         if (tx === undefined) {
             throw new NotFoundError();
         } else {
-            const resp = {
-                hash: tx.txHash,
-                status: tx.status,
-                gasPrice: tx.gasPrice,
-                updatedAt: tx.updatedAt,
-                blockNumber: tx.blockNumber,
-                expectedMinedInSec: tx.expectedMinedInSec,
-            };
-            res.status(HttpStatus.OK).send(resp);
+            res.status(HttpStatus.OK).send(marshallTransactionEntity(tx));
         }
     }
 }
@@ -244,4 +313,36 @@ const parseGetTransactionRequestParams = (req: express.Request): GetTransactionR
             ? undefined
             : parseUtils.parseStringArrForERC20BridgeSources(req.query.excludedSources.split(','));
     return { takerAddress, sellToken, buyToken, sellAmount, buyAmount, slippagePercentage, excludedSources };
+};
+
+interface PostTransactionRequestBody {
+    zeroExTransaction: ZeroExTransactionWithoutDomain;
+    signature: string;
+}
+
+const parsePostTransactionRequestBody = (req: any): PostTransactionRequestBody => {
+    const requestBody = req.body;
+    const signature = requestBody.signature;
+    const zeroExTransaction: ZeroExTransactionWithoutDomain = {
+        salt: new BigNumber(requestBody.zeroExTransaction.salt),
+        expirationTimeSeconds: new BigNumber(requestBody.zeroExTransaction.expirationTimeSeconds),
+        gasPrice: new BigNumber(requestBody.zeroExTransaction.gasPrice),
+        signerAddress: requestBody.zeroExTransaction.signerAddress,
+        data: requestBody.zeroExTransaction.data,
+    };
+    return {
+        zeroExTransaction,
+        signature,
+    };
+};
+
+const marshallTransactionEntity = (tx: TransactionEntity): any => {
+    return {
+        hash: tx.txHash,
+        status: tx.status,
+        gasPrice: tx.gasPrice,
+        updatedAt: tx.updatedAt,
+        blockNumber: tx.blockNumber,
+        expectedMinedInSec: tx.expectedMinedInSec,
+    };
 };
