+++ src/routers/meta_transaction_router.ts
@@ -24,5 +24,12 @@ export const createMetaTransactionRouter = (metaTransactionService: MetaTransact
      * GET status endpoint retrieves the transaction status by its hash.
      */
     router.get('/status/:txHash', asyncHandler(handlers.getTransactionStatusAsync.bind(handlers)));
+    /**
+     * POST Transaction endpoint takes a signed 0x Transaction and sends it to Ethereum
+     * for execution via `executeTransaction`.
+     *
+     * https://0x.org/docs/guides/v3-specification#executing-a-transaction
+     */
+    router.post('/submit', asyncHandler(handlers.submitZeroExTransactionIfWhitelistedAsync.bind(handlers)));
     return router;
 };
